# Documentation/api/openapi/weatherwear-api.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WeatherWear API",
    "description": "REST API for user authentication, weather lookup, AI clothing recommendations, recommendation history, feedback, and an AI style chat assistant.",
    "version": "1.0.0",
    "contact": {
      "name": "Dubrovin Mikita",
      "email": "dubrovin.nikita@student.ehu.lt"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8090/api",
      "description": "Local development server"
    },
    {
      "url": "https://wwtodipl-production.up.railway.app/api",
      "description": "Production server configured by application-prod.yml"
    }
  ],
...
  "paths": {
...
    "/history": {
      "get": {
        "tags": [
          "History"
        ],
        "operationId": "getHistory",
        "summary": "List current user's recommendation history",
        "description": "Returns recommendations generated by the authenticated user, ordered from newest to oldest.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Recommendation history list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryResponse"
                  }
                },
                "examples": {
                  "historyList": {
                    "summary": "Recommendation history list",
                    "value": [
                      {
                        "id": 101,
                        "city": "Vilnius",
                        "weatherSummary": "Temp: 12.4, Feels: 10.8, Condition: CLOUDS",
                        "recommendationText": "Wear a light waterproof jacket and closed shoes.",
                        "createdAt": "2026-04-30T12:45:00"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "History"
        ],
        "operationId": "clearHistory",
        "summary": "Clear current user's recommendation history",
        "description": "Deletes all recommendation history rows owned by the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "History cleared successfully. No response body is returned."
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/history/{id}": {
      "get": {
        "tags": [
          "History"
        ],
        "operationId": "getHistoryDetails",
        "summary": "Get one recommendation history item",
        "description": "Returns a single recommendation history item if it belongs to the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Recommendation history identifier.",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 101
          }
        ],
        "responses": {
          "200": {
            "description": "Recommendation history item.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                },
                "examples": {
                  "historyItem": {
                    "summary": "Recommendation history item",
                    "value": {
                      "id": 101,
                      "city": "Vilnius",
                      "weatherSummary": "Temp: 12.4, Feels: 10.8, Condition: CLOUDS",
                      "recommendationText": "Wear a light waterproof jacket and closed shoes.",
                      "createdAt": "2026-04-30T12:45:00"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT returned by /auth/register or /auth/login. Send it as Authorization: Bearer <token>."
      }
    },
    "responses": {
...
      "UnauthorizedError": {
        "description": "Authentication is missing, invalid, or expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "unauthorized": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 401,
                  "error": "Unauthorized",
                  "message": "Authentication is required",
                  "path": "/api/preferences"
                }
              }
            }
          }
        }
      },
...
      "NotFoundError": {
        "description": "The requested resource was not found or does not belong to the authenticated user.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "notFound": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 404,
                  "error": "Not Found",
                  "message": "Recommendation history not found",
                  "path": "/api/history/999"
                }
              }
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "serverError": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 500,
                  "error": "Internal Server Error",
                  "message": "Unexpected server error",
                  "path": "/api/recommendations"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "HistoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 101
          },
          "city": {
            "type": "string",
            "example": "Vilnius"
          },
          "weatherSummary": {
            "type": "string",
            "example": "Temp: 12.4, Feels: 10.8, Condition: CLOUDS"
          },
          "recommendationText": {
            "type": "string",
            "example": "Wear a light waterproof jacket and closed shoes."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-30T12:45:00"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "timestamp",
          "status",
          "error",
          "message",
          "path"
        ],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-30T12:45:00"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "email: Email must be valid"
          },
          "path": {
            "type": "string",
            "example": "/api/auth/register"
          }
        }
      },
...
    }
  }
}