# src/main/java/com/weatherwear/dto/history/HistoryResponse.java
package com.weatherwear.dto.history;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Getter;

import java.time.LocalDateTime;

@Getter
@AllArgsConstructor
@Schema(description = "Saved recommendation history item.")
public class HistoryResponse {

    @Schema(description = "Recommendation history identifier.", example = "101")
    private Long id;

    @Schema(description = "City used for the recommendation.", example = "Vilnius")
    private String city;

    @Schema(description = "Weather summary used by the recommendation prompt.",
            example = "Temp: 12.4, Feels: 10.8, Condition: CLOUDS")
    private String weatherSummary;

    @Schema(description = "Recommendation text generated by the AI provider.",
            example = "Wear a light waterproof jacket and closed shoes.")
    private String recommendationText;

    @Schema(description = "Timestamp when the recommendation was created.", example = "2026-04-30T12:45:00")
    private LocalDateTime createdAt;
}